# specs/001-review-comic-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PoSeeReview API
  description: |
    REST API for the PoSeeReview application - generates comic strips from strange restaurant reviews.
    
    **Key Features:**
    - Discover nearby restaurants using geolocation
    - Generate AI-powered comics from reviews
    - View real-time strangeness leaderboard
    
    **Tech Stack:** ASP.NET Core 9.0, Azure Table Storage, Azure OpenAI, DALL-E 3
  version: 1.0.0
  contact:
    name: PoSeeReview Team
  license:
    name: MIT
servers:
  - url: https://localhost:5001
    description: Local development (HTTPS)
  - url: http://localhost:5000
    description: Local development (HTTP)
  - url: https://poseereview.azurewebsites.net
    description: Production (Azure App Service)
...
paths:
...
  /api/comics/{placeId}:
    post:
      tags:
        - Comics
      summary: Generate comic for restaurant
      description: |
        Generates a four-panel comic strip from the restaurant's strangest reviews.
        Returns cached comic if generated within last 24 hours.
        
        **Generation Time:** 8-10 seconds (DALL-E API latency)
      operationId: generateComic
      parameters:
        - name: placeId
          in: path
          required: true
          description: Google Maps place ID
          schema:
            type: string
          example: "ChIJabcdef123"
        - name: forceRegenerate
          in: query
          required: false
          description: Force regeneration even if cache valid
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Comic generated successfully (or returned from cache)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicDto'
              example:
                comicId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                placeId: "ChIJabcdef123"
                restaurantName: "The Flying Saucer Café"
                strangenessScore: 87
                blobUrl: "https://poseereviews.blob.core.windows.net/comics/ChIJabcdef123/a1b2c3d4.png"
                generatedAt: "2025-01-22T14:35:00Z"
                expiresAt: "2025-01-23T14:35:00Z"
                isCached: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Comic generation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://tools.ietf.org/html/rfc7231#section-6.6.1"
                title: "Internal Server Error"
                status: 500
                detail: "DALL-E API failed: Rate limit exceeded"
                instance: "/api/comics/ChIJabcdef123"
                traceId: "00-abc123-def456-00"

    get:
      tags:
        - Comics
      summary: Get cached comic
      description: Retrieves the most recent cached comic for a restaurant (if exists and not expired).
      operationId: getCachedComic
      parameters:
        - name: placeId
          in: path
          required: true
          description: Google Maps place ID
          schema:
            type: string
          example: "ChIJabcdef123"
      responses:
        '200':
          description: Cached comic found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicDto'
        '404':
          description: No cached comic found (not generated or expired)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/leaderboard:
    get:
      tags:
        - Leaderboard
      summary: Get strangeness leaderboard
      description: |
        Returns top 10 restaurants ranked by highest strangeness score.
        Results are scoped to user's region for relevance.
      operationId: getLeaderboard
      parameters:
        - name: region
          in: query
          required: true
          description: Region code (e.g., US-CA-SF)
          schema:
            type: string
            pattern: '^[A-Z]{2}-[A-Z]{2}-[A-Z]+$'
          example: "US-WA-SEA"
        - name: limit
          in: query
          required: false
          description: Number of entries to return (default 10, max 50)
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Successfully retrieved leaderboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntryDto'
                  region:
                    type: string
                  lastUpdated:
                    type: string
                    format: date-time
              example:
                entries:
                  - rank: 1
                    placeId: "ChIJabcdef123"
                    restaurantName: "The Flying Saucer Café"
                    address: "123 Space Needle Way, Seattle, WA"
                    strangenessScore: 95
                    comicBlobUrl: "https://poseereviews.blob.core.windows.net/comics/ChIJabcdef123/xyz.png"
                  - rank: 2
                    placeId: "ChIJxyz789"
                    restaurantName: "Alien Burger Bar"
                    address: "456 Mars Ave, Seattle, WA"
                    strangenessScore: 87
                    comicBlobUrl: "https://poseereviews.blob.core.windows.net/comics/ChIJxyz789/abc.png"
                region: "US-WA-SEA"
                lastUpdated: "2025-01-22T14:40:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  schemas:
...
    ComicDto:
      type: object
      required:
        - comicId
        - placeId
        - restaurantName
        - strangenessScore
        - blobUrl
        - generatedAt
        - expiresAt
      properties:
        comicId:
          type: string
          format: uuid
        placeId:
          type: string
        restaurantName:
          type: string
        strangenessScore:
          type: number
          format: double
          minimum: 0
          maximum: 100
        blobUrl:
          type: string
          format: uri
          description: HTTPS URL to comic PNG in Azure Blob Storage
        generatedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          description: Cache expiration (generatedAt + 24 hours)
        isCached:
          type: boolean
          description: True if returned from cache, false if newly generated
    LeaderboardEntryDto:
      type: object
      required:
        - rank
        - placeId
        - restaurantName
        - strangenessScore
        - comicBlobUrl
      properties:
        rank:
          type: integer
          minimum: 1
          description: 1-based ranking
        placeId:
          type: string
        restaurantName:
          type: string
        address:
          type: string
        strangenessScore:
          type: number
          format: double
          minimum: 0
          maximum: 100
        comicBlobUrl:
          type: string
          format: uri
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
        traceId:
          type: string
          description: Correlation ID for logging (W3C Trace Context format)
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://tools.ietf.org/html/rfc7231#section-6.5.1"
            title: "Bad Request"
            status: 400
            detail: "Invalid latitude: must be between -90 and 90"
            instance: "/api/restaurants/nearby"
            traceId: "00-abc123-def456-00"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://tools.ietf.org/html/rfc7231#section-6.5.4"
            title: "Not Found"
            status: 404
            detail: "Restaurant with placeId 'ChIJinvalid' does not exist"
            instance: "/api/restaurants/ChIJinvalid"
            traceId: "00-xyz789-uvw123-00"
...
  securitySchemes: {}
security: []